fixed issue:(system.yaml file not found error due to initialization of member variable by code on kairon side)
